3.5

+++ trunk/R/runNonmem.R
@@ -2,7 +2,7 @@
 function (
 	run,
 	command,
-	ProjectDir,#where to make rundirs
+	ProjectDir,
 	boot,
 	urgent,
 	checkrunno,
@@ -27,27 +27,26 @@ function (
 	compile,
 	execute,
 	split,
-	outdir = ProjectDir,#where to put products
+	outdir = ProjectDir,
 	runext = if(boot) '.boot' else if(grid) '.lock' else '',
 	rundir = filename(ProjectDir,run,runext),
 	outfile = filename(rundir,run,'.lst'),
-	streams = ProjectDir,#where to find control streams
+	streams = ProjectDir,
 	ctlfile = filename(streams,run,'.ctl'),
-	tabfile = '(?<!par)\\.tab',
-	msffile = '^(?!\\$MSFI).*\\.msf',
-	parfile = 'par\\.tab',
 	remove  = c('^F[ISRC]','^OU','^nonmem.exe',if(fdata)c('^FD','^PR')),
 	...
 ){
   #Define some functions.
   final <- function(x)sub('\\.lock','',x)
-  scavenge <- function(expr,lines)lines[grep(expr,lines,ignore.case=TRUE, perl=TRUE)][[1]]
-  lock <- function(x)Sys.chmod(x,mode='0664')
   
   #Groom arguments.
   outdir <- star(outdir,run)
   rundir <- star(rundir,run)
   ctlfile <- star(ctlfile,run)
+  outfile <- star(outfile,run)
+  tabfile <- tabfile(ctlfile,dir=final(rundir),...)
+  parfile <- parfile(ctlfile,dir=final(rundir),...)
+  msffile <- msffile(ctlfile,dir=final(rundir),...)
   if(!file.exists(ctlfile))message(paste(ctlfile,'was not found'))
   if(!file.exists(ctlfile))return()
   control <- explicitPath(readLines(ctlfile))
@@ -59,24 +58,22 @@ function (
 	  script <- epilog
 	  epilog <- episcript
   }
-  parfile <- resolve(scavenge(parfile,control),rundir)
-  tabfile <- resolve(scavenge(tabfile,control),rundir)
-  msffile <- resolve(scavenge(msffile,control),rundir)
   
   #Prepare the file environment.
-  if(!is.null(file))if(file.exists(file))file.remove(file)
-  if(file.exists(outfile))file.remove(outfile)
-  if(file.exists(tabfile))file.remove(tabfile)
-  if(file.exists(parfile))file.remove(parfile)
-  if(file.exists(msffile))file.remove(msffile)
-  purge.dir(final(rundir),nice)
-  if(grid)purge.dir(rundir)
-  dir.create(rundir, showWarnings = FALSE)
-  dname <- getdname(ctlfile)
-  if(!file.exists(resolve(dname,rundir)))warning(paste(dname,'not visible from',rundir))
-  if(!file.exists(resolve(dname,rundir)))return()
-  file.copy(ctlfile, file.path(rundir,basename(ctlfile)), overwrite = TRUE)
-
+  if(compile){
+  	  if(!is.null(file))if(file.exists(file))file.remove(file)
+	  if(file.exists(outfile))file.remove(outfile)
+	  if(file.exists(tabfile))file.remove(tabfile)
+	  if(file.exists(parfile))file.remove(parfile)
+	  if(file.exists(msffile))file.remove(msffile)
+	  purge.dir(final(rundir),nice)
+	  if(rundir!=final(rundir))purge.dir(rundir)
+	  dir.create(rundir, showWarnings = FALSE)
+	  dname <- getdname(ctlfile)
+	  if(!file.exists(resolve(dname,rundir)))warning(paste(dname,'not visible from',rundir))
+	  if(!file.exists(resolve(dname,rundir)))return()
+	  file.copy(ctlfile, file.path(rundir,basename(ctlfile)), overwrite = TRUE)
+  }
   #Run NONMEM.
   runCommand(
   	command=command,
@@ -97,74 +94,64 @@ function (
   )
 
   #Clean up.
-  if(boot & grid)return() #boot runs on grid have sync=n by definition, so run perh. not complete.
-  lapply(remove,purge.files,dir=rundir)
-  #lock(dir(rundir,pattern=paste('^',run,'\\.',sep=''),full.names=TRUE))
-  #lock(dir(rundir,pattern='n.*\\.',full.names=TRUE))
-  runfilename <- function(rundir)dir(rundir,'^Run',full.names=TRUE)
-  try(file.rename(runfilename(rundir),sub('\\.o[0-9]*$','.out',runfilename(rundir))),silent=TRUE)
-  if(grid){
-  	dir.create(final(rundir), showWarnings = FALSE)
-  	file.copy(from=dir(rundir,full.names=TRUE),to=final(rundir),overwrite=TRUE)
-   	purge.dir(rundir)
+  if(execute){
+	  if(boot & grid)return() #boot runs on grid have sync=n by definition, so run perh. not complete.
+	  lapply(remove,purge.files,dir=rundir)
+	  if(rundir!=final(rundir)){
+		dir.create(final(rundir), showWarnings = FALSE)
+		file.copy(from=dir(rundir,full.names=TRUE),to=final(rundir),overwrite=TRUE)
+		purge.dir(rundir)
+	  }
+	
+	  #Diagnostics
+	  if(!udef)
+	   if(nmVersion(config(dirname(command))) < 7)
+	    try(setCwres(cwres=getCwres(directory=final(rundir)),file=tabfile))
+	  if(diag)try(
+		PLOTR(
+			run,
+			ProjectDir,
+			dvname,
+			logtrans,
+			grp,
+			grpnames,
+			cont.cov,
+			cat.cov,
+			par.list,
+			eta.list,
+			missing,
+			file=file,
+			ctlfile=ctlfile,
+			outfile=final(outfile),
+			rundir=final(rundir),
+			outdir=outdir,
+			...
+		)
+	  )
+	  if (!is.null(epilog))if(is.function(epilog))try(
+		  epilog(
+			run=run,
+			ProjectDir=ProjectDir,
+			dvname=dvname,
+			logtrans=logtrans,
+			grp=grp,
+			grpames=grpnames,
+			cont.cov=cont.cov,
+			cat.cov=cat.cov,
+			par.list=par.list,
+			eta.list=eta.list,
+			missing=missing,
+			ctlfile=ctlfile,
+			outfile=final(outfile),
+			rundir=final(rundir),
+			outdir=outdir,
+			...,
+			script=script
+		)
+	  )
+	
+	  message(paste("Run ", run, " complete.", sep = ""))
   }
-  #lock(dir(final(rundir)))
-  #lock(tabfile)
-  #lock(parfile)
-  #lock(msffile)
-
-  #Diagnostics
-  if(!udef)
-   if(nmVersion(config(dirname(command))) < 7)
-    try(setCwres(cwres=getCwres(directory=final(rundir)),file=tabfile))
-  if(diag)try(
-    	PLOTR(
-    		run,
-		ProjectDir,
-		dvname,
-		logtrans,
-		grp,
-		grpnames,
-		cont.cov,
-		cat.cov,
-		par.list,
-		eta.list,
-		missing,
-		file=file,
-		tabfile=tabfile,
-		ctlfile=ctlfile,
-		parfile=parfile,
-		outfile=final(outfile),
-		rundir=final(rundir),
-		outdir=outdir,
-		...
-        )
-  )
-  if (!is.null(epilog))if(is.function(epilog))try(
-	  epilog(
-		run=run,
-		ProjectDir=ProjectDir,
-		dvname=dvname,
-		logtrans=logtrans,
-		grp=grp,
-		grpames=grpnames,
-		cont.cov=cont.cov,
-		cat.cov=cat.cov,
-		par.list=par.list,
-		eta.list=eta.list,
-		missing=missing,
-		tabfile=tabfile,
-		ctlfile=ctlfile,
-		parfile=parfile,
-		outfile=final(outfile),
-		rundir=final(rundir),
-		outdir=outdir,
-		...,
-		script=script
-	)
-  )
-
-  message(paste("Run ", run, " complete.", sep = ""))
 }
 
 #.............................................................................
@@ -213,9 +200,11 @@ function (
   }
   extractPath <- function(x)sub('(^.*(MSFO?|FILE) *= *)([^ ]*)(.*$)','\\3',x,ignore.case=TRUE)
   resolve <- function(file,dir)ifelse(contains('^\\.',file),file.path(dir,file),file)
+  scavenge <- function(expr,lines)lines[grep(expr,lines,ignore.case=TRUE, perl=TRUE)][[1]]
 	  
-	  
-	  
+  tabfile <- function(ctlfile,dir,tabreg='(?<!par)\\.tab',...)resolve(scavenge(tabreg,explicitPath(readLines(ctlfile))),dir)
+  parfile <- function(ctlfile,dir,parreg='par\\.tab',...)resolve(scavenge(parreg,explicitPath(readLines(ctlfile))),dir)
+  msffile <- function(ctlfile,dir,msfreg='^(?!\\$MSFI).*\\.msf',...)resolve(scavenge(msfreg,explicitPath(readLines(ctlfile))),dir)
